NO-JIRA: c++: Enable -Werror -pedantic -Weverything for clang++ compiler, fix warnings.
- #if with undefined macro
- old style casts (converted to enum conversions and const_cast)
- sign/precision in implicit conversions
- unused parameters
- redundant default and break statements.
- drop noexcept/throw()

Following warnings are disabled, some may be worth enabling: -Wno-c++98-compat -Wno-c++98-compat-pedantic -Wno-float-equal -Wno-padded -Wno-sign-conversion -Wno-switch-enum -Wno-weak-vtables -Wno-exit-time-destructors -Wno-global-constructors -Wno-shorten-64-to-32

+++ proton-c/bindings/cpp/src/data.cpp
@@ -87,7 +87,7 @@ int compare_container(data& a, data& b) {
     int cmp = compare(sa.is_described, sb.is_described);
     if (cmp) return cmp;
     // Lexical sort (including descriptor if there is one)
-    size_t min_size = std::min(sa.size, sb.size) + int(sa.is_described);
+    size_t min_size = std::min(sa.size, sb.size) + size_t(sa.is_described);
     for (size_t i = 0; i < min_size; ++i) {
         cmp = compare_next(a, b);
         if (cmp) return cmp;
