+++ proton-c/bindings/cpp/src/proton_bits.cpp
@@ -23,7 +23,7 @@
 #include <proton/object.h>
 #include "proton_bits.hpp"
 
-std::string error_str(int code) {
+std::string error_str(long code) {
   switch (code)
   {
   case 0: return "ok";
@@ -39,7 +39,7 @@ std::string error_str(int code) {
   }
 }
 
-std::string error_str(pn_error_t* err, int code) {
+std::string error_str(pn_error_t* err, long code) {
     if (err && pn_error_code(err)) {
         const char* text = pn_error_text(err);
         return text ? std::string(text) : error_str(pn_error_code(err));
